Added Auth to codes - updated controllers - updated tests - assimilated sub-codes tests in codes tests
