+++ src/controllers/CourtRoleCodesController.ts
@@ -2,8 +2,10 @@ import { Body, Delete, Get, Path, Post, Put, Query, Route } from 'tsoa';
 import { CourtRoleCode } from '../models/CourtRoleCode';
 import { CourtRoleCodeService } from '../services/CourtRoleCodeService';
 import ControllerBase from './ControllerBase';
+import { Security } from '../authentication';
 
 @Route('codes/courtroles')
+@Security('jwt')
 export class CourtRoleCodesController extends ControllerBase<CourtRoleCode> {
 
     get service(){
