+++ src/controllers/LeaveCancelCodesController.ts
@@ -2,8 +2,10 @@ import { Get, Route } from 'tsoa';
 import ControllerBase from './ControllerBase';
 import { LeaveCancelReasonCode } from '../models/LeaveCancelReasonCode';
 import { LeaveCancelReasonCodeService } from '../services/LeaveCancelCodeService';
+import { Security } from '../authentication';
 
 @Route('codes/leave-cancel')
+@Security('jwt')
 export class LeaveCancelCodesController extends ControllerBase<LeaveCancelReasonCode> {
 
     get service() {
