+++ src/controllers/LeaveSubTypeCodesController.ts
@@ -2,8 +2,10 @@ import { Get, Route } from 'tsoa';
 import ControllerBase from './ControllerBase';
 import { LeaveSubCode } from '../models/LeaveSubCode';
 import { LeaveSubCodeService } from '../services/LeaveSubCodeService';
+import { Security } from '../authentication';
 
 @Route('codes/leave-sub-type')
+@Security('jwt')
 export class LeaveSubTypeCodesController extends ControllerBase<LeaveSubCode> {
 
     get service() {
