+++ src/controllers/OtherAssignCodesController.ts
@@ -2,8 +2,10 @@ import { Body, Delete, Get, Path, Post, Put, Query, Route } from 'tsoa';
 import { OtherAssignCode } from '../models/OtherAssignCode';
 import { OtherAssignCodeService } from '../services/OtherAssignCodeService';
 import ControllerBase from './ControllerBase';
+import { Security } from '../authentication';
 
 @Route('codes/otherassign')
+@Security('jwt')
 export class OtherAssignCodesController extends ControllerBase<OtherAssignCode> {
 
     get service(){
